+++ CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 
 4.5.1, 2021-02-12
 -----------------
+- ask for confirmation when exporting HTML with media in subfolder, but login unavailable for downloading media (issue #93).
 - fixed not retaining the order of command-line FILE arguments (issue #93);
 - add detected media type extension to shared media exported in HTML if lacking one (issue #93).